==> scripts/02_continental_divide.Rmd <==
---
title: "Western North America Boundary"
author: "Jason Pither"
date: "2024-01-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

NOTE from Hannah: I was unable to save the files using relative paths. I had to 
input my working directory in order for this .Rmd to run. I don't know how to 
adjust the working directory properly for this file without messing up everything else.


## Background

This script aims to create a geospatial vector dataset that provides the boundary for the continental divide of North America, buffered by 200km to its east, then bounded by the Arctic and Pacific oceans, and the Mexico border.

## Load libraries

```{r}
library(sf)
library(dplyr)
library(tmap)
library(ggplot2)
```

## Data sources

We use hydrological basins to delimit the continental divide. Basins are acquired from this ["hydrosheds" website](https://www.hydrosheds.org/products/hydrobasins#downloads) the "standard" versions of the Arctic and North America datasets.

Unzip those, and you'll see a PDF document to read.

## Import data

We want the level02 shapefiles, for both arctic and north america

```{r}
basin.na <- st_read("C:/Users/PilatH/OneDrive - AGR-AGR/Desktop/ranunculus_sdm/data/raw/hybas_na_lev01-12_v1c/hybas_na_lev02_v1c.shp")
basin.ar <- st_read("C:/Users/PilatH/OneDrive - AGR-AGR/Desktop/ranunculus_sdm/data/raw/hybas_ar_lev01-12_v1c/hybas_ar_lev02_v1c.shp")
```
Note that many of the variables are being treated as numeric, when they should be character.  The key field is "HYBAS_ID", which needs to be converted to character.


```{r}
basin.ar$HYBAS_ID <- as.character(basin.ar$HYBAS_ID)
basin.na$HYBAS_ID <- as.character(basin.na$HYBAS_ID)
```

Let's see the unique values

```{r}
basin.ar$HYBAS_ID
```

```{r}
basin.na$HYBAS_ID
```

We'll use these codes to filter out the basins of interest

## Subset to focal basins

First, look at the maps of the coarsest level basins, to get the codes needed for the basins of interest.

It turns out the features in the vector shapefiles (polygons) are pretty detailed, and we don't need this.  We can simplify to speed up processing.

For info, see this [chapter](https://r.geocompx.org/geometry-operations.html?q=st_simplify#simplification).

Note that we first get an error:

>Error in wk_handle.wk_wkb(wkb, s2_geography_writer(oriented = oriented,  : 
  Loop 542 is not valid: Edge 7860 has duplicate vertex with edge 7874

So to fix this, we use the `st_make_valid` function... of course we do!  then we can simplify.

But let's first check the spatial reference system on these layers:

```{r}
st_crs(basin.na)
```

OK, WGS84 datum, with length units as metres.  

So, when simplifying borders use a "tolerance" of 100m

First Arctic:

```{r}
basin.ar <- st_make_valid(basin.ar)
arctic.simple <- st_simplify(basin.ar, dTolerance = 100)
```

Do same for North America

```{r}
basin.na <- st_make_valid(basin.na)
na.simple <- st_simplify(basin.na, dTolerance = 100)
```

Now we can map them to isolate desired basins:

```{r}
tm_shape(arctic.simple) + tm_fill(fill = "HYBAS_ID")
```

OK, so ID# 8020000010 is the one we want

Now North America:

```{r}
tm_shape(na.simple) + tm_fill(fill = "HYBAS_ID")
```

ID# 7020000010, and 7020014250

Now we can  filter to the basins of interest, then merge.

IMPORTANT: here you'll see that i include the `dplyr::filter` approach, to ensure that the correct `filter` function is used.

```{r}
arctic.filtered <- arctic.simple |>
  dplyr::filter(HYBAS_ID == "8020000010")
na.filtered <- na.simple |>
  dplyr::filter(HYBAS_ID == "7020000010" | HYBAS_ID == "7020014250")
```

Now verify with maps

```{r}
tm_shape(arctic.filtered) + tm_fill(fill = "HYBAS_ID")
```


```{r}
tm_shape(na.filtered) + tm_fill(fill = "HYBAS_ID")
```

Now merge the two maps

```{r}
western.divide <- rbind(arctic.filtered, na.filtered)
```

And dissolve the field to make one big polygon.

```{r}
western.divide <- st_union(western.divide, by_feature = FALSE)
```

```{r}
ggplot(western.divide) + geom_sf(fill = "grey")
```

Hmm, it looks like there's borders still among basins, which we don't want...

Let's re-do the dissolving, first with a buffer of 10m

```{r}
western.divide <- rbind(arctic.filtered, na.filtered)
western.divide <- western.divide |>
  st_buffer(10) |>
  st_union()
```

```{r}
ggplot(western.divide) + geom_sf(fill = "grey")
```

OK! that looks useful.

## Write to file

Now we should export this layer, then we can create a new buffered layer too.

```{r}
write_sf(western.divide, "C:/Users/PilatH/OneDrive - AGR-AGR/Desktop/ranunculus_sdm/data/extents/continental_divide.shp")
```

## Buffer eastward

Now we can buffer the entire polygon, because we don't care about it expanding westward into the ocean, because that won't have any practical effect when overlaying with occurrence records.

Let's buffer 200km (200,000m). Note we again need to use `st_make_valid` first, because we initially get an error...

```{r}
western.divide.buffer <- western.divide |>
  st_make_valid() |>
  st_buffer(200000) |>
  st_sf()
```

```{r}
ggplot(western.divide.buffer) + geom_sf(fill = "grey")
```

oops, that's cool.. it spills into the other hemisphere!

let's try 100km

```{r}
western.divide.buffer <- western.divide |>
  st_make_valid() |>
  st_buffer(100000) |>
  st_sf()
```

```{r}
ggplot(western.divide.buffer) + geom_sf(fill = "grey")
```

NOPE.

OK, so what we'll need to do is make a new dataset that includes the basins just to the east of these ones, dissolve the polygones, then buffer those borders, then use those buffered borders to clip the eastern-most border of the focal (western) basins, then do the buffer at that point.

the relevant book chapter is [here](https://r.geocompx.org/geometry-operations#clipping).

## Isolate basins east of Rockies

Let's go back to the maps

```{r}
tm_shape(arctic.simple) + tm_fill(fill = "HYBAS_ID")
```

So the ID# of interest is:  8020008900

```{r}
tm_shape(na.simple) + tm_fill(fill = "HYBAS_ID")
```

So the ID# of interest are:  7020021430, 7020046750, 7020047840

Filter:

```{r}
arctic.inner.filtered <- arctic.simple |>
  dplyr::filter(HYBAS_ID == "8020008900")
na.inner.filtered <- na.simple |>
  dplyr::filter(HYBAS_ID == "7020021430" | HYBAS_ID == "7020046750" | HYBAS_ID == "7020047840")
```

Merge

```{r}
western.inner.divide <- rbind(arctic.inner.filtered, na.inner.filtered)
western.inner.divide <- western.inner.divide |>
  st_buffer(10) |>
  st_union()
```

```{r}
ggplot(western.inner.divide) + geom_sf(fill = "grey") 
  
```

OK, now let's see the overlap with our western polygon

**NOTE** to see this you should copy and paste the code in the console, run it there.
```{r}
plot(western.inner.divide) +
  plot(western.divide, add = TRUE) 
```

Perfect, that gives us what we want.  we now can use `st_intersection` to get what we need.

```{r}
western.border <- st_intersection(st_make_valid(western.divide), st_make_valid(western.inner.divide))
```

**NOTE** to see this you should copy and paste the code in the console, run it there.
```{r}
plot(western.inner.divide, col = "green") +
  plot(western.divide, col = "grey", add = TRUE) +
  plot(western.border, col = "red", add = TRUE)
```

Now we can buffer the border of the divide 200km eastward!

```{r}
western.border.buffer <- western.border |>
  st_make_valid() |>
  st_buffer(200000)
```

**NOTE** to see this you should copy and paste the code in the console, run it there.
```{r}
plot(western.inner.divide, col = "green") +
  plot(western.divide, col = "grey", add = TRUE) +
  plot(western.border.buffer, col = "red", add = TRUE)
```

great!

Now we can dissolve the buffer region with the western polygon

```{r}
western.final.boundary <- st_union(st_make_valid(st_sf(western.divide)), st_make_valid(st_sf(western.border.buffer)))
```

```{r}
ggplot(western.final.boundary) + geom_sf(fill = "grey")
```

## Write to file

```{r}
write_sf(western.final.boundary, "C:/Users/PilatH/OneDrive - AGR-AGR/Desktop/ranunculus_sdm/data/extents/continental_divide_buffer_boundary.shp")
```
